core/remote: Move DirectoryNotFound to errors
  We group all remote error classes in the `errors` sub-module make
  finding them easier and avoid circular dependencies in other modules.
