+++ test/unit/remote/index.js
@@ -14,7 +14,7 @@ const metadata = require('../../../core/metadata')
 const Prep = require('../../../core/prep')
 const remote = require('../../../core/remote')
 const { Remote } = remote
-const { DirectoryNotFound } = require('../../../core/remote/cozy')
+const { DirectoryNotFound } = require('../../../core/remote/errors')
 const { ROOT_DIR_ID, TRASH_DIR_ID } = require('../../../core/remote/constants')
 const { FetchError } = require('../../../core/remote/cozy')
 const timestamp = require('../../../core/utils/timestamp')
